Page breaks ===========
It seems that you can't have it all.
If you have a heading with a "page-break-after: avoid", and that the
paragraph after has a "page-break-inside: avoid", the latter will be
applied, but the heading won't come along on the new page.
You can then insert manual page breaks, either by setting a "page-break"
class to the heading, or add an empty <div> with a "page-break" class
where you want the new page to begin.